pyproject.toml
[tool.poetry]
name = "Artist Management System"
version = "0.1.0"
description = "REST API For Artist Management System"
authors = ["Sijan Shrestha <sijan.shr9z@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.2.0"
argon2-cffi = "^23.1.0"
django = "^5.0.3"
django-environ = "^0.11.2"
django-model-utils = "^4.4.0"
django-allauth = "^0.61.1"
django-crispy-forms = "^2.1"
crispy-bootstrap5 = "^2024.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
drf-spectacular = "^0.27.1"
psycopg = {extras = ["c"], version = "^3.1.18"}
django-rest-knox = "^4.2.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.7.1"
werkzeug = {extras = ["watchdog"], version = "^3.0.1"}
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.7"}
pytest = "^8.0.2"
pytest-sugar = "^1.0.0"
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.14.5"}
ruff = "^0.3.1"
djlint = "^1.34.1"
pre-commit = "^3.6.2"
django-debug-toolbar = "^4.3.0"
django-extensions = "^3.2.3"
pytest-django = "^4.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
